infra/requirements.txt
aws-cdk-lib==2.175.1
constructs>=10.0.0
